Custom Backend Task
We move the custom-backend-task.ts file to the root directory of our project.  We change the Data type alias/data structure to a record with a single field called hello of type String, which means it can contain a text value returned by the hello(name) function in custom-backend-task.ts.  We change the body of the data function and the body of the view function to display app.data.hello.